Image slider
let slideIndex = 1;
This variable holds the index of the currently displayed slide. It's initialized to 1, indicating that the first slide should be displayed initially.

showSlides(slideIndex);
This line calls the showSlides function to display the slide specified by the slideIndex.

Function plusSlides(n)
This function is called when the next or previous arrow buttons are clicked. It takes a parameter n which can be either 1 (next) or -1 (previous). It increments or decrements the slideIndex by n and then calls the showSlides function to display the corresponding slide.

Function currentSlide(n)
This function is called when one of the thumbnail dots is clicked. It directly sets the slideIndex to the value of the parameter n and then calls the showSlides function to display the corresponding slide.

Function showSlides(n)
This is the main function that manages the display of slides and controls their visibility.

a. It starts by declaring local variables i, slides, and dots. slides holds an array of elements with the class "mySlides", which represent the individual slides, and dots holds an array of elements with the class "dot", which represent the thumbnail dots for navigation.

b. The function first checks if n is greater than the total number of slides. If n is greater, it wraps around to the first slide by setting slideIndex to 1.

c. Similarly, it checks if n is less than 1. If so, it wraps around to the last slide by setting slideIndex to the total number of slides.

d. The function then iterates through all slides, hiding them by setting their display CSS property to "none".

e. It also iterates through all the thumbnail dots, removing the "active" class from each of them.

f. After hiding all slides and removing the "active" class from dots, it displays the slide corresponding to the updated slideIndex by setting its display CSS property to "block".

g. It also adds the "active" class to the corresponding thumbnail dot, marking it as the currently active slide.